requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
pandas==2.3.2
pyarrow==21.0.0
python-dotenv==1.1.1
google-genai==1.29.0
pytest==8.4.2
